Update from Melchior Franz
- translates objects to English language
- adds tail rotor protector
- adds exhausts
- adds tail handles
- fixes model bugs (missing blade twister levers, ...)
- makes glasses a little less blue
- welds fuselage halves together
- fixes pilot position
- disables notorque
- fixes rear door anim (/controls/doors/rear: 0 closed, 1 open)

It also sets the magneto switch to zero. Especially since
the joystick's throttle input (collective) is reversed (as
it should be), it may happen that you have full collective
when you last flew an airplane. And I like the startup
sound. But you may set the magneto to 1, if you want.  :-)