Melchior FRANZ:
Not a patch, but a drop-in file. It has all objects parented under object
"Aircraft", which shouldn't be displayed in fgrun. This makes the bo105 now
actually visable with fgrun (from cvs/head). Not a raisin any more!

PS: This ac file was filtered through my new "AC3D sorter, checker & sanitizer":
    see http://members.aon.at/mfranz/flightgear/